[Chore] #1 ErrorStatus 등 패키지 구조 수정
