+++ src/main/java/com/nova/yeobaek/domain/closet/status/ClosetErrorStatus.java
@@ -1,15 +1,15 @@
-package com.nova.yeobaek.global.payload.code;
+package com.nova.yeobaek.domain.closet.status;
 
 import org.springframework.http.HttpStatus;
 
-import com.nova.yeobaek.global.payload.ErrorReason;
+import com.nova.yeobaek.global.payload.status.ErrorReason;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum OOTDErrorStatus implements ErrorReason {
+public enum ClosetErrorStatus implements ErrorReason {
 
 	;
