+++ src/main/java/com/nova/yeobaek/domain/ootd/status/OOTDErrorStatus.java
@@ -1,15 +1,15 @@
-package com.nova.yeobaek.global.payload.code;
+package com.nova.yeobaek.domain.ootd.status;
 
 import org.springframework.http.HttpStatus;
 
-import com.nova.yeobaek.global.payload.ErrorReason;
+import com.nova.yeobaek.global.payload.status.ErrorReason;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public enum ItemErrorStatus implements ErrorReason {
+public enum OOTDErrorStatus implements ErrorReason {
 
 	;
