+++ src/main/java/com/nova/yeobaek/global/payload/status/SuccessStatus.java
@@ -1,4 +1,4 @@
-package com.nova.yeobaek.global.payload.common;
+package com.nova.yeobaek.global.payload.status;
 
 import org.springframework.http.HttpStatus;
